refactor(indexer): the block cutter uploads the block it cut
`BlockUploader` was one object per database, threaded from `Database.kt` through `LogProcessor` into each term's `BlockCutter`, and it had exactly one caller: `BlockCutter.upload`, which discarded the message id it returned. Nothing about it was per-database except its lifetime, and that lifetime was the wrong one — closing a block is a term's act, under a term's fence, stamped with a term's id. So `uploadBlock` becomes a private method on the cutter, and a parameter disappears from three constructors.

- **`BlockUploaded` was the one leader-side replica write that went round the append seam.**
  It was a direct `replicaLog.appendMessage`, which made it unreachable from `LogProcessor.LogsDriver` — a test could stall or fail every other append and not that one. It now goes through the driver like everything else, so the cutter holds both seams and every leader-side append is injectable.

- **It stays out of the append pump, though, and that asymmetry with the boundary beside it is the point.**
  Queueing it would drop the happens-before between the append and `liveIndex.nextBlock()`. The pump's `shutdown` discards whatever is queued, so a term ending in that gap leaves this node committed to the block — block files written, catalog refreshed, index rolled — with nothing on the log to release the followers buffering behind the boundary, and no way to re-cut it, because `openFollower` carries no pending block and its own catalog has already moved on. Recovery would need leadership to reach a node still holding the pending block.
  Losing a queued *boundary* is benign by comparison: nothing anywhere has advanced, and finishing an unfinished block is exactly what the next leader's inherited-pending-block path is for. The promotion path makes the difference concrete — it uploads the inherited block before `replicaAppender.run()` is launched inside `runTerm`, so a queued message there would sit with no consumer at all for the length of `replayHeldRecords`.
  The awaited append is the narrow fix. The wider one is to stop applying the upload eagerly at all and let the leader roll its own index when it reads its own `BlockUploaded` back, the way `FollowerLogProcessor.closeBlock` already does — which would make the ordering structural rather than enforced, and turn a lost message into overwritable orphan files. That changes the promotion state machine, so it is not this commit.

- **The metrics stay the database's, because a gauge keeps the state object it was registered with.**
  `Gauge.builder("xtdb.block.last_upload_time", lastUploadEpochSeconds)` hands Micrometer an `AtomicLong`, and a second registration under the same name and tags returns the first meter and drops the new state. Registered per term, the gauge would therefore report the first term's last upload for the life of the process — which is precisely the reading it exists to contradict, since it is there to expose a database whose blocks have quietly stopped landing (#5867). So `LogProcessor` registers it once and hands each term's cutter the `AtomicLong` behind it.

- **`gcDispatcher` and `uploadDispatcher` were one thing under two names.**
  Both existed so a simulation could seed them, and `LogProcessorSimTest` passed the same dispatcher to both; they differed only in that the upload one arrived pre-capped by `limitedParallelism`. Capping at the consumer is what the rest of the codebase already does — `Compactor` derives `jobsDispatcher` off its `ioDispatcher`, and each GC derives a table and a delete view off its own — so `BlockCutter` now derives the 16-way upload cap itself and `LogProcessor` carries a single `ioDispatcher`. `LiveIndex.open` already had the name.
  One thing this does give up: the deleted top-level `defaultBlockUploadDispatcher` was a single 16-way view for the whole JVM, where the cutter's is one per term per database, and a `limitedParallelism` view of `Dispatchers.IO` is elastic rather than drawing on shared permits. So the in-flight `TableBlock.toByteArray()` buffers are bounded per database now, not per node.

- **The term argument went too, without changing what is stamped.**
  `uploadBlock` took `uploadingTermId` separately from the boundary, and the only caller passed the cutter's own term. A promotion still uploads the previous leader's block under the incoming term, because the cutter doing it is the incoming term's — the fact worth keeping, which is now a comment at the `termId` it sets rather than a parameter nobody varied.

- **Test churn is all deletion.**
  Fourteen `BlockUploader` constructions go from `LogProcessorTest` along with the helper parameter that carried them; `LeaderTermTest`'s two harnesses and `ExternalSourceTest`'s build the cutter directly. The eleven affected classes — 83 tests across the leader, follower, term-fence, block and live-index suites — pass unchanged.

- **`db.allium`'s `BlockUploader` entity becomes `BlockCutter`'s.**
  The spec had a full entity for the uploader and named the cutter only in a `See:` line, which had the ownership backwards. The seven-step contract moves under `BlockCutter` with the block cycle it already owned, the db-system component list loses `block_uploader`, and the promotion and boundary rules call `log_processor.block_cutter.upload(...)`. `block-gc.allium`'s cross-reference follows the rename. `allium check` reports the same diagnostics as before the edit, and no findings.

Refs #5917
