--- core/src/main/kotlin/xtdb/indexer/BlockUploader.kt
@@ -1,179 +1,0 @@
-package xtdb.indexer
-
-import io.micrometer.core.instrument.Gauge
-import io.micrometer.core.instrument.MeterRegistry
-import io.micrometer.core.instrument.Timer
-import java.nio.ByteBuffer
-import java.time.Instant
-import java.util.concurrent.atomic.AtomicLong
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers.IO
-import kotlinx.coroutines.coroutineScope
-import kotlinx.coroutines.launch
-import xtdb.types.MessageId
-import xtdb.api.log.ReplicaMessage
-import xtdb.api.log.ReplicaMessage.BlockBoundary
-import xtdb.api.log.ReplicaMessage.BlockUploaded
-import xtdb.api.log.SourceMessage
-import xtdb.api.storage.Storage
-import xtdb.catalog.TableCatalog
-import xtdb.compactor.Compactor
-import xtdb.database.Database
-import xtdb.database.PartitionState
-import xtdb.database.PartitionStorage
-import xtdb.log.proto.TrieDetails
-import xtdb.util.StringUtil.asLexHex
-import xtdb.util.debug
-import xtdb.util.logger
-
-private val LOG = BlockUploader::class.logger
-
-// Cap concurrent per-table block-metadata uploads to bound the heap held by in-flight
-// `TableBlock.toByteArray()` buffers. S3's default HTTP client caps concurrency at 50, so
-// parallelism beyond this would queue inside the SDK with the `byte[]`s still pinned on heap.
-private const val MAX_CONCURRENT_BLOCK_UPLOADS = 16
-private val defaultBlockUploadDispatcher = IO.limitedParallelism(MAX_CONCURRENT_BLOCK_UPLOADS, "block-upload")
-
-class BlockUploader(
-    partitionStorage: PartitionStorage,
-    partitionState: PartitionState,
-    dbName: String,
-    private val compactor: Compactor.ForDatabase,
-    private val dbCatalog: Database.Catalog?,
-    private val meterRegistry: MeterRegistry?,
-    private val scope: CoroutineScope,
-    private val uploadDispatcher: CoroutineDispatcher = defaultBlockUploadDispatcher,
-) {
-    private val sourceLog = partitionStorage.sourceLog
-    private val replicaLog = partitionStorage.replicaLog
-    private val bufferPool = partitionStorage.bufferPool
-    private val liveIndex = partitionState.liveIndex
-    private val trieCatalog = partitionState.trieCatalog
-    private val tableCatalog = partitionState.tableCatalog
-    private val blockUploadTimer: Timer? = meterRegistry?.let {
-        Timer.builder("block.upload.timer")
-            .publishPercentiles(0.75, 0.85, 0.95, 0.98, 0.99, 0.999)
-            .register(it)
-    }
-
-    // A timer records uploads that happened; this records the absence of one. An external source
-    // confirms its upstream position only as far as the last durable block, so a database whose blocks
-    // have quietly stopped landing pins the upstream's log while ingestion, queries and healthz all stay
-    // green — time-since-last-block is what makes that visible (#5867).
-    private val lastUploadEpochSeconds = AtomicLong(0)
-
-    init {
-        meterRegistry?.let { reg ->
-            Gauge.builder("xtdb.block.last_upload_time", lastUploadEpochSeconds) { it.get().toDouble() }
-                .description("epoch seconds at which this database's most recent block landed in object storage")
-                .baseUnit("seconds")
-                .tag("db", dbName)
-                .register(reg)
-        }
-    }
-
-    /**
-     * Snapshot the live index into block files, append [boundary]'s matching `BlockUploaded`, and roll the
-     * index. Returns the `BlockUploaded`'s replica-log position.
-     *
-     * [uploadingTermId] is not always [boundary]'s: a transition finishes the previous leader's pending
-     * block, and the `BlockUploaded` carries the term of whoever appends it, because a leader confirms a
-     * write only on reading it back at its own term.
-     */
-    suspend fun uploadBlock(
-        boundaryReplicaMsgId: MessageId, uploadingTermId: Long,
-        boundary: BlockBoundary,
-    ): MessageId {
-        val latestProcessedMsgId = boundary.latestProcessedMsgId
-        val blockIdx = boundary.blockIndex
-        LOG.debug("finishing block: 'b${blockIdx.asLexHex}'...")
-        val timer = meterRegistry?.let { Timer.start(it) }
-
-        val finishedBlocks = liveIndex.finishBlock(bufferPool, blockIdx)
-
-        // A table staged with no rows has no trie — see LiveTable.FinishedBlock.writtenTrie. It still
-        // reaches the table catalog below, so its declared columns survive.
-        val addedTries =
-            finishedBlocks.mapNotNull { (table, fb) ->
-                val writtenTrie = fb.writtenTrie ?: return@mapNotNull null
-
-                val trieDetails = TrieDetails.newBuilder()
-                    .setTableName(table.schemaAndTable)
-                    .setTrieKey(writtenTrie.trieKey)
-                    .setDataFileSize(writtenTrie.dataFileSize)
-                    .also { it.setTrieMetadata(writtenTrie.trieMetadata) }
-                    .build()
-
-                // NOTE: side-effect here.
-                trieCatalog.addTries(table, listOf(trieDetails), Instant.now())
-
-                trieDetails
-            }
-
-        val allTables = finishedBlocks.keys + tableCatalog.allTables
-        val tablePartitions = allTables.associateWith { trieCatalog.getPartitions(it) }
-
-        val tableBlocks = tableCatalog.finishBlock(finishedBlocks, tablePartitions)
-
-        // A table new in this block has already written its L0 trie by now, under the slug its LiveTable was
-        // created with — the same one minted here, because both resolve through `State.slug`.
-        val entries = tableCatalog.resolveTables(tableBlocks.keys).associateBy { it.table }
-
-        coroutineScope {
-            tableBlocks.forEach { (table, tableBlock) ->
-                launch(uploadDispatcher) {
-                    val path = TableCatalog.tableBlockPath(entries.getValue(table).slug, blockIdx)
-                    bufferPool.putObject(path, ByteBuffer.wrap(tableBlock.toByteArray()))
-                }
-            }
-        }
-
-        val secondaryDatabasesForBlock = dbCatalog?.serialisedSecondaryDatabases
-
-        val externalSourceToken = boundary.externalSourceToken
-
-        val block = tableCatalog.buildBlock(
-            blockIdx, liveIndex.latestCompletedTx, latestProcessedMsgId,
-            boundaryReplicaMsgId, entries.values, secondaryDatabasesForBlock,
-            externalSourceToken, uploadingTermId
-        )
-
-        bufferPool.putObject(TableCatalog.blockFilePath(blockIdx), ByteBuffer.wrap(block.toByteArray()))
-        tableCatalog.refresh(block)
-        lastUploadEpochSeconds.set(Instant.now().epochSecond)
-
-        // Now signal followers that the block is available.
-        val uploadedMsgId = replicaLog.appendMessage(
-            BlockUploaded(
-                Storage.VERSION, bufferPool.epoch,
-                blockIdx, latestProcessedMsgId,
-                addedTries, externalSourceToken,
-                termId = uploadingTermId,
-            )
-        ).msgId
-
-        LOG.debug("block uploaded b${blockIdx.asLexHex}: source=$latestProcessedMsgId, replica=$uploadedMsgId")
-
-        blockUploadTimer?.let { timer?.stop(it) }
-        liveIndex.nextBlock()
-
-        // Publish L0 tries to the source log so that all nodes — including multi-writer nodes running
-        // concurrently with a single-writer leader — see the L0 before any compaction L1C on the source
-        // log (see #5395). Once we drop support for multi-writer clusters running concurrently with
-        // single-writer, this source-log post is no longer required and can be removed.
-        //
-        // Both off the persister coroutine, in one launch: uploadBlock runs on the source log's sole
-        // consumer, so appending inline self-deadlocks when the source-log buffer saturates at a block
-        // boundary — the consumer would wait on room only it can make. The launch lets uploadBlock return
-        // so the persister drains and the append's emit finds room; running signalBlock after it in the
-        // same coroutine keeps compaction's L1C strictly behind the L0 without a separate join.
-        scope.launch {
-            sourceLog.appendMessage(SourceMessage.TriesAdded(Storage.VERSION, bufferPool.epoch, addedTries))
-            compactor.signalBlock()
-        }
-        LOG.debug("finished block: 'b${blockIdx.asLexHex}'.")
-
-        return uploadedMsgId
-    }
-}
