+++ dev/doc/high-level-tour.adoc
@@ -35,10 +35,10 @@ State - in-memory state holders::
 `TableCatalog` (`xtdb.catalog`), `TrieCatalog` (interface in `xtdb.trie`, implemented in `xtdb.trie-catalog`), and `LiveIndex` (`xtdb.indexer`).
 
 Services - orchestration::
-`LogProcessor` and its three role implementations, `BlockUploader` and `CrashLogger` (`xtdb.indexer`); `Compactor` (`xtdb.compactor`); `BlockGarbageCollector` and `TrieGarbageCollector` (`xtdb.garbage-collector`); and `DatabaseCatalog` (`xtdb.database` - a service despite the name, managing database attach/detach and lifecycle).
+`LogProcessor` and its three role implementations, `BlockCutter` and `CrashLogger` (`xtdb.indexer`); `Compactor` (`xtdb.compactor`); `BlockGarbageCollector` and `TrieGarbageCollector` (`xtdb.garbage-collector`); and `DatabaseCatalog` (`xtdb.database` - a service despite the name, managing database attach/detach and lifecycle).
 
 Dependencies flow inward: services depend on state and on storage; state depends on neither.
-State components *yield* data and the service does the I/O - `LiveIndex.finishBlock` takes a `BufferPool` as a call parameter rather than holding one, and it's `BlockUploader` (which does hold one) that supplies it and writes the resulting files.
+State components *yield* data and the service does the I/O - `LiveIndex.finishBlock` takes a `BufferPool` as a call parameter rather than holding one, and it's `BlockCutter` (which does hold one) that supplies it and writes the resulting files.
 That's 'functional core, imperative shell' in OO vocabulary.
 
 `PartitionState` and `PartitionStorage` (`xtdb.database`) state the split most explicitly: the former bundles the three state holders, the latter the buffer pool, the source/replica logs and the page-metadata factory.
@@ -46,7 +46,7 @@ That's 'functional core, imperative shell' in OO vocabulary.
 Two caveats, so you don't read more into this than is there:
 
 * *It isn't a package boundary.*
-  `xtdb.catalog`, `xtdb.compactor` and `xtdb.garbage-collector` are clean, but `xtdb.indexer` holds `LiveIndex` (state) alongside the log processors and `BlockUploader` (services), and `xtdb.trie` holds the `TrieCatalog` interface alongside the trie/data/metadata file writers.
+  `xtdb.catalog`, `xtdb.compactor` and `xtdb.garbage-collector` are clean, but `xtdb.indexer` holds `LiveIndex` (state) alongside the log processors and `BlockCutter` (services), and `xtdb.trie` holds the `TrieCatalog` interface alongside the trie/data/metadata file writers.
 * *`TableCatalog` keeps a `BufferPool` for its lifetime*, unlike `TrieCatalog` (whose I/O sits in the factory).
   It's only used by `loadTables`, called once from `PartitionState.open` at startup - so 'state does no runtime I/O' holds here by convention rather than by construction.
